Set program to reset after five rounds
One thing I hope to learn in the future is how to compartmentalize
functions so that I don't cram all my functions under one
function like how I did on this one. I think this is due to my lack of
understanding in DOM manipulation. Also, I hope to understand someday
how to make it so that my alert comes after the fifth round is displayed
on screen. The result of the fifth round is displayed after the alert.
The player won't know the result of the fifth round until they've seen
the result of the whole game first. But I still am pround of myself for
being able to make this simple game using only javascript and html
regardless of how un-user friendly it is.